requirements.txt
opencv-python
numpy
matplotlib
torchvision
fastapi
uvicorn
python-multipart
jinja2
pillow
uvicorn[standard]

